+++ features/support/ruby_features.rb
@@ -27,3 +27,13 @@ Around "@skip-when-required-keyword-args-unsupported" do |scenario, block|
     warn "Skipping scenario #{scenario.title} because required keyword arguments are not supported"
   end
 end
+
+Around "@skip-when-ripper-unsupported" do |scenario, block|
+  require 'rspec/support/ruby_features'
+
+  if ::RSpec::Support::RubyFeatures.ripper_supported?
+    block.call
+  else
+    warn "Skipping scenario #{scenario.title} because Ripper is not supported"
+  end
+end